Cleaning up and testing
